implemented login check
