+++ routes/api.php
@@ -1,10 +1,13 @@
 <?php
 
-use App\Http\Controllers\Api\Register\Plain;
+use App\Http\Controllers\Api\Plain;
 use Illuminate\Support\Facades\Route;
 
 Route::namespace('Api')->group(function () {
     Route::name('register.')->prefix('register')->namespace('Register')->group(function () {
         Route::post('/plain', [Plain::class, 'register'])->name('plain');
+    });
+    Route::name('login.')->prefix('login')->namespace('Login')->group(function () {
+        Route::post('/plain', [Plain::class, 'login'])->name('plain');
     });
 });
